+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [1.0.0] - decode and isExpired methods
+## [1.0.0] - decode() and isExpired() methods
 
 - decode method will decode your token's payload and return a Map<String, dynamic>
 - isExpired method is used to know if a token is already expired or not
@@ -8,3 +8,8 @@
 - An example file was added
 
 ## [1.0.2] - License was added
+
+## [1.0.3] - getExpirationDate() and getTokenTime() methods
+
+- getExpirationDate() returns the expiration date of the token
+- getTokenTime() you can use this method to know how old your token is